chore: tests adjustment
